+++ packages/framer-motion/src/components/AnimatePresence/index.tsx
@@ -16,6 +16,7 @@ import { PresenceChild } from "./PresenceChild"
 import { LayoutGroupContext } from "../../context/LayoutGroupContext"
 import { useIsomorphicLayoutEffect } from "../../utils/use-isomorphic-effect"
 import { useUnmountEffect } from "../../utils/use-unmount-effect"
+import { warnOnce } from "../../utils/warn-once"
 
 type ComponentKey = string | number
 
@@ -75,14 +76,23 @@ function onlyElements(children: ReactNode): ReactElement<any>[] {
  *
  * @public
  */
-export const AnimatePresence: React.FunctionComponent<React.PropsWithChildren<AnimatePresenceProps>> = ({
+export const AnimatePresence: React.FunctionComponent<
+    React.PropsWithChildren<AnimatePresenceProps>
+> = ({
     children,
     custom,
     initial = true,
     onExitComplete,
     exitBeforeEnter,
     presenceAffectsLayout = true,
+    mode = "sync",
 }) => {
+    // Support deprecated exitBeforeEnter prop
+    if (exitBeforeEnter) {
+        mode = "wait"
+        warnOnce(false, "Replace exitBeforeEnter with mode='wait'")
+    }
+
     // We want to force a re-render once all exiting animations have finished. We
     // either use a local forceRender function, or one from a parent context if it exists.
     let [forceRender] = useForceUpdate()
@@ -132,6 +142,7 @@ export const AnimatePresence: React.FunctionComponent<React.PropsWithChildren<An
                         isPresent
                         initial={initial ? undefined : false}
                         presenceAffectsLayout={presenceAffectsLayout}
+                        mode={mode}
                     >
                         {child}
                     </PresenceChild>
@@ -160,7 +171,7 @@ export const AnimatePresence: React.FunctionComponent<React.PropsWithChildren<An
 
     // If we currently have exiting children, and we're deferring rendering incoming children
     // until after all current children have exiting, empty the childrenToRender array
-    if (exitBeforeEnter && exiting.size) {
+    if (mode === "wait" && exiting.size) {
         childrenToRender = []
     }
 
@@ -205,6 +216,7 @@ export const AnimatePresence: React.FunctionComponent<React.PropsWithChildren<An
                 onExitComplete={onExit}
                 custom={custom}
                 presenceAffectsLayout={presenceAffectsLayout}
+                mode={mode}
             >
                 {child}
             </PresenceChild>
@@ -222,6 +234,7 @@ export const AnimatePresence: React.FunctionComponent<React.PropsWithChildren<An
                 key={getChildKey(child)}
                 isPresent
                 presenceAffectsLayout={presenceAffectsLayout}
+                mode={mode}
             >
                 {child}
             </PresenceChild>
@@ -230,11 +243,11 @@ export const AnimatePresence: React.FunctionComponent<React.PropsWithChildren<An
 
     if (
         env !== "production" &&
-        exitBeforeEnter &&
+        mode === "wait" &&
         childrenToRender.length > 1
     ) {
         console.warn(
-            `You're attempting to animate multiple children within AnimatePresence, but its exitBeforeEnter prop is set to true. This will lead to odd visual behaviour.`
+            `You're attempting to animate multiple children within AnimatePresence, but its mode is set to "wait". This will lead to odd visual behaviour.`
         )
     }